+++ nix/modules/project-modules.nix
@@ -16,44 +16,21 @@ in
               A lazy attrset of `haskellProjects.<name>` modules that can be
               imported in other flakes.
             '';
-            defaultText = ''
+            defaultText = lib.literalMD ''
               Package and dependency information for this project exposed for reuse
               in another flake, when using this project as a Haskell dependency.
 
-              Typically the consumer of this flake will want to use one of the
-              following modules:
-
-                - output: provides both local package and dependency overrides.
-                - local: provides only local package overrides (ignores dependency
-                  overrides in this flake)
-
-              These default modules are always available.
+              The 'output' module of the default project is included by default,
+              returning `defaults.projectModules.output`.
             '';
-            default = { }; # Set in config (see ./default-project-modules.nix)
+            default = { };
           };
 
-        config.haskellFlakeProjectModules =
-          let
-            defaults = rec {
-              # The 'output' module provides both local package and dependency
-              # overrides.
-              output = {
-                imports = [ input local ];
-              };
-              # The 'local' module provides only local package overrides.
-              local = { pkgs, lib, ... }: withSystem pkgs.system ({ config, ... }: {
-                source-overrides =
-                  lib.mapAttrs (_: v: v.root)
-                    config.haskellProjects.default.packages;
-              });
-              # The 'input' module contains only dependency overrides.
-              input = { pkgs, ... }: withSystem pkgs.system ({ config, ... }: {
-                inherit (config.haskellProjects.default)
-                  source-overrides overrides;
-              });
-            };
-          in
-          defaults;
+        config.haskellFlakeProjectModules = {
+          output = { pkgs, lib, ... }: withSystem pkgs.system ({ config, ... }:
+            config.haskellProjects."default".defaults.projectModules.output
+          );
+        };
       }];
     };
   };