+++ test/simple/flake.nix
@@ -19,11 +19,19 @@
         inputs.haskell-flake.flakeModule
         inputs.check-flake.flakeModule
       ];
-      flake.haskellFlakeProjectModules.default = { pkgs, ... }: {
-        overrides = self: super: {
+      flake.haskellFlakeProjectModules.default = { pkgs, lib, ... }: {
+        packages = {
           # This is purposefully incorrect (pointing to ./.) because we
           # expect it to be overriden in perSystem below.
-          foo = self.callCabal2nix "foo" ./. { };
+          foo.source = ./.;
+        };
+        settings = {
+          # Test that self and super are passed
+          foo = { self, super, ... }: {
+            custom = _: builtins.seq 
+              (lib.assertMsg (lib.hasAttr "ghc" self) "self is bad") 
+              super.foo;
+          };
         };
         devShell = {
           tools = hp: {
@@ -33,15 +41,16 @@
           hlsCheck.enable = true;
         };
       };
-      perSystem = { self', pkgs, ... }: {
+      perSystem = { self', pkgs, lib, ... }: {
         haskellProjects.default = {
           # Multiple modules should be merged correctly.
           imports = [ self.haskellFlakeProjectModules.default ];
-          overrides = self: super: {
-            # This overrides the overlay above (in `flake.*`), because the
-            # module system merges them in such order. cf. the WARNING in option
-            # docs.
-            foo = self.callCabal2nix "foo" (inputs.haskell-multi-nix + /foo) { };
+          # Debug logging should work.
+          debug = true;
+          packages = {
+            # Because the module being imported above also defines a root for
+            # the 'foo' package, we must override it here using `lib.mkForce`.
+            foo.source = lib.mkForce (inputs.haskell-multi-nix + /foo);
           };
           devShell = {
             tools = hp: {