+++ test/simple/test.sh
@@ -1,9 +1,18 @@
 source ../common.sh
 set -euxo pipefail
 
+rm -f result result-bin
+
 # First, build the flake
 logHeader "Testing nix build"
 nix build ${OVERRIDE_ALL}
+
+# Test defaults.settings module behaviour, viz: separateBinOutput
+test -f result-bin/bin/haskell-flake-test || {
+    echo "ERROR: separateBinOutput (from defaults.settings) not in effect"
+    exit 1
+}
+
 # Run the devshell test script in a nix develop shell.
 logHeader "Testing nix devshell"
 nix develop ${OVERRIDE_ALL} -c ./test-in-devshell.sh
@@ -13,5 +22,6 @@ nix run ${OVERRIDE_ALL} .#app1
 # Test non-devshell features:
 # Checks
 logHeader "Testing nix flake checks"
-nix --option sandbox false \
-    build ${OVERRIDE_ALL} -L .#check
+nix --option sandbox false build \
+    --no-link --print-out-paths \
+    ${OVERRIDE_ALL} -L .#check